requirements.txt
fastapi
uvicorn
pydantic
pymongo[srv]
python-dotenv
mongoengine
passlib
bcrypt
pytest
httpx
python-multipart
python-jose[cryptography]
PyJWT